+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 #0.0.4
 
 * Basic OAuth stubbing works
+* forked from `xolvio:github-fake`